requirements.txt
requests==2.28.1
azure-iot-device==2.12.0
